refactor: Position 패키지 이동
